Begin route refactor to update tabs routing and move to route const files.
