+++ src/app/pages/about/about-routes.ts
@@ -1,0 +1,9 @@
+import { Routes } from '@angular/router';
+import { AboutPage } from './about';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: AboutPage
+  }
+];
