+++ src/app/pages/account/account-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { AccountPage } from './account';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: AccountPage
+  }
+];
