+++ src/app/pages/login/login-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { LoginPage } from './login';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: LoginPage
+  }
+];
