+++ src/app/pages/login/login.module.ts
@@ -2,16 +2,17 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
+import { RouterModule } from '@angular/router';
 
 import { LoginPage } from './login';
-import { LoginPageRoutingModule } from './login-routing.module';
+import { routes } from './login-routes';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    LoginPageRoutingModule
+    RouterModule.forChild(routes)
   ],
   declarations: [
     LoginPage,
