+++ src/app/pages/map/map-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { MapPage } from './map';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: MapPage
+  }
+];
