+++ src/app/pages/map/map.module.ts
@@ -1,15 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
+import { RouterModule } from '@angular/router';
 
 import { MapPage } from './map';
-import { MapPageRoutingModule } from './map-routing.module';
+import { routes } from './map-routes';
 
 @NgModule({
   imports: [
     CommonModule,
     IonicModule,
-    MapPageRoutingModule
+    RouterModule.forChild(routes)
   ],
   declarations: [
     MapPage,
