--- src/app/pages/schedule/schedule-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { SchedulePage } from './schedule';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: SchedulePage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class SchedulePageRoutingModule { }
