+++ src/app/pages/session-detail/session-detail-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { SessionDetailPage } from './session-detail';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: SessionDetailPage
+  }
+];
