+++ src/app/pages/signup/signup-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { SignupPage } from './signup';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: SignupPage
+  }
+];
