+++ src/app/pages/signup/signup.module.ts
@@ -2,16 +2,17 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
+import { RouterModule } from '@angular/router';
 
 import { SignupPage } from './signup';
-import { SignupPageRoutingModule } from './signup-routing.module';
+import { routes } from './signup-routes';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    SignupPageRoutingModule
+    RouterModule.forChild(routes)
   ],
   declarations: [
     SignupPage,
