+++ src/app/pages/speaker-detail/speaker-detail-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { SpeakerDetailPage } from './speaker-detail';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: SpeakerDetailPage
+  }
+];
