--- src/app/pages/speaker-list/speaker-list-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { SpeakerListPage } from './speaker-list';
-const routes: Routes = [
-  {
-    path: '',
-    component: SpeakerListPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class SpeakerListPageRoutingModule {}
