+++ src/app/pages/speaker-list/speaker-list.module.ts
@@ -1,15 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
+import { RouterModule } from '@angular/router';
 
 import { SpeakerListPage } from './speaker-list';
-import { SpeakerListPageRoutingModule } from './speaker-list-routing.module';
+import { routes } from './speaker-list-routes';
 
 @NgModule({
   imports: [
     CommonModule,
     IonicModule,
-    SpeakerListPageRoutingModule
+    RouterModule.forChild(routes)
   ],
   declarations: [SpeakerListPage],
 })
