+++ src/app/pages/support/support-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { SupportPage } from './support';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: SupportPage
+  }
+];
