--- src/app/pages/support/support-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { SupportPage } from './support';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: SupportPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class SupportPageRoutingModule { }
