+++ src/app/pages/support/support.module.ts
@@ -2,15 +2,17 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
+import { RouterModule } from '@angular/router';
+
 import { SupportPage } from './support';
-import { SupportPageRoutingModule } from './support-routing.module';
+import { routes } from './support-routes';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    SupportPageRoutingModule
+    RouterModule.forChild(routes)
   ],
   declarations: [
     SupportPage,
