+++ src/app/pages/tabs-page/tabs-page-routes.ts
@@ -1,0 +1,41 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { TabsPage } from './tabs-page';
+
+
+const routes: Routes = [
+  {
+    path: 'tabs',
+    component: TabsPage,
+    children: [
+      {
+        path: 'schedule',
+        loadChildren: () => import('../schedule/schedule.module').then(m => m.ScheduleModule)
+      },
+      {
+        path: 'speakers',
+        loadChildren: () => import('../speaker-list/speaker-list.module').then(m => m.SpeakerListModule)
+      },
+      {
+        path: 'map',
+        loadChildren: () => import('../map/map.module').then(m => m.MapModule)
+      },
+      {
+        path: 'about',
+        loadChildren: () => import('../about/about.module').then(m => m.AboutModule)
+      },
+      {
+        path: '',
+        redirectTo: '/app/tabs/schedule',
+        pathMatch: 'full'
+      }
+    ]
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class TabsPageRoutingModule { }
+
