+++ src/app/pages/tutorial/tutorial-routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { TutorialPage } from './tutorial';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: TutorialPage
+  }
+];
