--- src/app/pages/tutorial/tutorial-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { TutorialPage } from './tutorial';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: TutorialPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class TutorialPageRoutingModule { }
